did some refactoring, cleaned out main.py left to do config saving and loading
